コンポーネントとhooksに切り分け
