+++ client/src/components/organisms/aboutMe/EditMe.tsx
@@ -1,50 +1,25 @@
 import { memo, FC } from "react";
-import { useForm } from "react-hook-form";
-import * as yup from "yup";
-import { yupResolver } from "@hookform/resolvers/yup";
-
-//バリデーションチェック
-const schema = yup.object().shape({
-  //姓のバリデーション
-  name: yup
-    .string()
-    .trim()
-    .required("姓名を入力してください")
-    .max(15, "姓名は15文字以内で入力してください"),
-});
+import { TextInput } from "../../atoms/editMe/TextInput";
+import { useUserInfo } from "../../../hooks/editMe/useUserInfo";
 
 /**
  * EditMe.
  */
 export const EditMe: FC = memo(() => {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm({
-    resolver: yupResolver(schema),
-    defaultValues: {
-      name: "名前",
-    },
-  });
-  /**
-   * 更新ボタンを押した時に呼ばれる
-   * @param data - 入力したデータ
-   */
-  const onSubmit = async (data: any) => {
-    try {
-      alert("データ:" + data.name);
-    } catch (error) {
-      console.log(error);
-    }
-  };
+  const { handleSubmit, cancel, register, errors, onSubmit } =
+    useUserInfo("初期表示用データ");
 
   return (
     <>
-      {errors.name?.message}
-      <input type="text" {...register("name")} />
+      <TextInput
+        registers={register("name")}
+        errorMessage={errors.name?.message}
+      />
       <button type="button" onClick={handleSubmit(onSubmit)}>
         送信
+      </button>
+      <button type="button" onClick={cancel}>
+        キャンセル
       </button>
     </>
   );
